+++ stable_baselines/common/callbacks.py
@@ -1,13 +1,17 @@
 import os
 from abc import ABC, abstractmethod
+import typing
 from typing import Union, List, Dict, Any, Optional
 
 import gym
 import numpy as np
 
-from stable_baselines.common.base_class import BaseRLModel # pytype: disable=pyi-error
 from stable_baselines.common.vec_env import VecEnv, sync_envs_normalization
 from stable_baselines.common.evaluation import evaluate_policy
+from stable_baselines import logger
+
+if typing.TYPE_CHECKING:
+    from stable_baselines.common.base_class import BaseRLModel  # pytype: disable=pyi-error
 
 
 class BaseCallback(ABC):
@@ -25,17 +29,20 @@ class BaseCallback(ABC):
         self.verbose = verbose
         self.locals = None  # type: Dict[str, Any]
         self.globals = None  # type: Dict[str, Any]
+        self.logger = None  # type: logger.Logger
         # Sometimes, for event callback, it is useful
         # to have access to the parent object
         self.parent = None  # type: Optional[BaseCallback]
 
-    def init_callback(self, model: BaseRLModel) -> None:
+    # Type hint as string to avoid circular import
+    def init_callback(self, model: 'BaseRLModel') -> None:
         """
         Initialize the callback by saving references to the
         RL model and the training environment for convenience.
         """
         self.model = model
         self.training_env = model.get_env()
+        self.logger = logger.Logger.CURRENT
         self._init_callback()
 
     def _init_callback(self) -> None:
@@ -50,16 +57,15 @@ class BaseCallback(ABC):
     def _on_training_start(self) -> None:
         pass
 
-    # Should we include that?
-    # def on_rollout_start(self, locals_: Dict[str, Any], globals_: Dict[str, Any]) -> None:
-    #     pass
+    def on_rollout_start(self) -> None:
+        self._on_rollout_start()
+
+    def _on_rollout_start(self) -> None:
+        pass
 
     @abstractmethod
     def _on_step(self) -> bool:
         """
-        TODO: Should we modify current implementation?
-        i.e. call after each env step instead after each rollout (current implementation)?
-
         :return: (bool) If the callback returns False, training is aborted early.
         """
         return True
@@ -81,9 +87,11 @@ class BaseCallback(ABC):
     def _on_training_end(self) -> None:
         pass
 
-    # Should we include that?
-    # def on_rollout_end(self, locals_: Dict[str, Any], globals_: Dict[str, Any]) -> None:
-    #     pass
+    def on_rollout_end(self) -> None:
+        self._on_rollout_end()
+
+    def _on_rollout_end(self) -> None:
+        pass
 
 
 class EventCallback(BaseCallback):
@@ -101,16 +109,21 @@ class EventCallback(BaseCallback):
         if callback is not None:
             self.callback.parent = self
 
-    def init_callback(self, model: BaseRLModel) -> None:
+    def init_callback(self, model: 'BaseRLModel') -> None:
         super(EventCallback, self).init_callback(model)
-        self.callback.init_callback(self.model)
+        if self.callback is not None:
+            self.callback.init_callback(self.model)
 
     def _on_training_start(self) -> None:
-        self.callback.on_training_start(self.locals, self.globals)
+        if self.callback is not None:
+            self.callback.on_training_start(self.locals, self.globals)
 
     def _on_event(self) -> bool:
         if self.callback is not None:
             return self.callback()
+        return True
+
+    def _on_step(self) -> bool:
         return True
 
 
@@ -128,6 +141,10 @@ class CallbackList(BaseCallback):
         for callback in self.callbacks:
             callback.on_training_start(self.locals, self.globals)
 
+    def _on_rollout_start(self) -> None:
+        for callback in self.callbacks:
+            callback.on_rollout_start()
+
     def _on_step(self) -> bool:
         continue_training = True
         for callback in self.callbacks:
@@ -137,6 +154,10 @@ class CallbackList(BaseCallback):
             # Return False (stop training) if at least one callback returns False
             continue_training = callback() and continue_training
         return continue_training
+
+    def _on_rollout_end(self) -> None:
+        for callback in self.callbacks:
+            callback.on_rollout_end()
 
     def _on_training_end(self) -> None:
         for callback in self.callbacks:
@@ -175,7 +196,7 @@ class ConvertCallback(BaseCallback):
     """
     Convert functional callback (old-style) to object.
 
-    :param on_step: (callable)
+    :param callback: (Callable)
     :param verbose: (int)
     """
     def __init__(self, callback, verbose=0):
@@ -197,12 +218,13 @@ class EvalCallback(EventCallback):
         when there is a new best model according to the `mean_reward`
     :param n_eval_episodes: (int) The number of episodes to test the agent
     :param eval_freq: (int) Evaluate the agent every eval_freq call of the callback.
-    :param log_path: (str) Path to a log file (.npz) where the evaluations
+    :param log_path: (str) Path to a folder where the evaluations (`evaluations.npz`)
         will be saved. It will be updated at each evaluation.
     :param best_model_save_path: (str) Path to a folder where the best model
         according to performance on the eval env will be saved.
     :param deterministic: (bool) Whether the evaluation should
         use a stochastic or deterministic actions.
+    :param render: (bool) Whether to render or not the environment during evaluation
     :param verbose: (int)
     """
     def __init__(self, eval_env: Union[gym.Env, VecEnv],
@@ -212,20 +234,27 @@ class EvalCallback(EventCallback):
                  log_path: str = None,
                  best_model_save_path: str = None,
                  deterministic: bool = True,
+                 render: bool = False,
                  verbose: int = 1):
         super(EvalCallback, self).__init__(callback_on_new_best, verbose=verbose)
         self.n_eval_episodes = n_eval_episodes
         self.eval_freq = eval_freq
         self.best_mean_reward = -np.inf
         self.deterministic = deterministic
+        self.render = render
+
         if isinstance(eval_env, VecEnv):
             assert eval_env.num_envs == 1, "You must pass only one environment for evaluation"
 
         self.eval_env = eval_env
         self.best_model_save_path = best_model_save_path
+        # Logs will be written in `evaluations.npz`
+        if log_path is not None:
+            os.path.join(log_path, 'evaluations')
         self.log_path = log_path
         self.evaluations_results = []
         self.evaluations_timesteps = []
+        self.evaluations_length = []
 
     def _init_callback(self):
         # Does not work when eval_env is a gym.Env and training_env is a VecEnv
@@ -240,22 +269,30 @@ class EvalCallback(EventCallback):
 
     def _on_step(self) -> bool:
 
-        if self.n_calls % self.eval_freq == 0:
+        if self.eval_freq > 0 and self.n_calls % self.eval_freq == 0:
             # Sync training and eval env if there is VecNormalize
             sync_envs_normalization(self.training_env, self.eval_env)
 
-            episode_rewards, _ = evaluate_policy(self.model, self.eval_env, n_eval_episodes=self.n_eval_episodes,
-                                                 deterministic=self.deterministic, return_episode_rewards=True)
+            episode_rewards, episode_lengths = evaluate_policy(self.model, self.eval_env,
+                                                               n_eval_episodes=self.n_eval_episodes,
+                                                               render=self.render,
+                                                               deterministic=self.deterministic,
+                                                               return_episode_rewards=True)
 
             if self.log_path is not None:
                 self.evaluations_timesteps.append(self.num_timesteps)
                 self.evaluations_results.append(episode_rewards)
-                np.savez(self.log_path, timesteps=self.evaluations_timesteps, results=self.evaluations_results)
+                self.evaluations_length.append(episode_lengths)
+                np.savez(self.log_path, timesteps=self.evaluations_timesteps,
+                         results=self.evaluations_results, ep_lengths=self.evaluations_length)
 
             mean_reward, std_reward = np.mean(episode_rewards), np.std(episode_rewards)
+            mean_ep_length, std_ep_length = np.mean(episode_lengths), np.std(episode_lengths)
+
             if self.verbose > 0:
                 print("Eval num_timesteps={}, "
                       "episode_reward={:.2f} +/- {:.2f}".format(self.num_timesteps, mean_reward, std_reward))
+                print("Episode length: {:.2f} +/- {:.2f}".format(mean_ep_length, std_ep_length))
 
             if mean_reward > self.best_mean_reward:
                 if self.verbose > 0:
@@ -291,8 +328,8 @@ class StopTrainingOnRewardThreshold(BaseCallback):
         # Convert np.bool to bool, otherwise callback() is False won't work
         continue_training = bool(self.parent.best_mean_reward < self.reward_threshold)
         if self.verbose > 0 and not continue_training:
-            print("Stopping training because the mean reward {} is above the threshold "
-                 "{}".format(self.parent.best_mean_reward, self.reward_threshold))
+            print("Stopping training because the mean reward {:.2f} "
+                  " is above the threshold {}".format(self.parent.best_mean_reward, self.reward_threshold))
         return continue_training